+++ sistem_rs_be/src/migrations/20250707064720_create_angsuran_table.js
@@ -9,10 +9,10 @@ export const up = function (knex) {
       .references('IDINVOICE').inTable('invoice').onDelete('CASCADE');
     table.string('NIK', 20).notNullable()
       .references('NIK').inTable('pasien').onDelete('CASCADE');
-    table.decimal('JUMLAHTOTAL', 15, 2).notNullable();
-    table.decimal('JUMLAH_TERBAYAR', 15, 2).defaultTo(0);
-    table.decimal('SISA_ANGSURAN', 15, 2).notNullable();
-    table.decimal('JUMLAH_CICILAN', 15, 2).notNullable();
+    table.double('JUMLAHTOTAL', 15, 2).notNullable();
+    table.double('JUMLAH_TERBAYAR', 15, 2).defaultTo(0);
+    table.double('SISA_ANGSURAN', 15, 2).notNullable();
+    table.double('JUMLAH_CICILAN', 15, 2).notNullable();
     table.datetime('TANGGAL_JATUH_TEMPO').notNullable();
     table.enu('STATUS', ['AKTIF', 'LUNAS', 'MACET']).defaultTo('AKTIF');
     table.timestamp('CREATED_AT').defaultTo(knex.fn.now());
